==> 04_partycmte_dash.Rmd <==
---
title: "DEM VS. GOP PARTY COMMITTTEES"
output: 
  flexdashboard::flex_dashboard:
    orientation: rows
    vertical_layout: fill
---

```{r setup, include=FALSE}

library(flexdashboard)
library(tidyverse)
library(lubridate)
library(janitor)
library(glue)
library(plotly)
library(DT)
library(googlesheets)
library(kableExtra)
library(leaflet)
# library(ggmap)
library(RColorBrewer)
library(htmlwidgets)
library(htmltools)
library(tigris)
options(tigris_class = "sf")


### DATA LOADING AND PREP

#bring in saved table from step 01
zipcompare <- readRDS("zipcompare.rds")

#bring in chart objects from step 02
#save as RDS object
top10chart_dem <- readRDS("demtopzip_plt_nm.rds")
top10chart_gop <- readRDS("goptopzip_plt_nm.rds")

# load saved map object from step 03
savedmap_zipcompare <- readRDS("savedmap_zipcompare.rds")


```


Table & Top 10 Charts
===================================== 
Breakdown by zip code as reported in the January monthly FEC reports filed on Feb 20.  
This dataset includes individual contributions to the **DNC/DNCC** vs. the **RNC/NRCC**.

Row {data-height=650}
-----------------------------------------------------------------------

### {data-width=800}

```{r}

mytable <- zipcompare %>% 
  select(zip_code, city, state, everything(), -zipname) %>% 
  filter(!state %in% c("VI", "PR", "AE"),
         !zip_code %in% c("0", "00000", "99999", "SW1V"),
         !is.na(city)) %>% 
  mutate(
    state = as.factor(state),
    winner = as.factor(winner)
  )




datatable(mytable, 
          rownames = FALSE, 
          filter = "top",
          options = list(searchHighlight = TRUE)) %>% 
  formatCurrency(c('demtotal', 'goptotal', 'advantage'), digits = 0) %>%   
  formatStyle('demtotal', color = 'darkblue') %>% 
  formatStyle('goptotal', color = 'darkred') %>% 
  formatStyle('zip_code', fontWeight = "bold")





```



Row {data-height=350}
-----------------------------------------------------------------------

### 

```{r}

top10chart_dem


```


### 

```{r}

top10chart_gop


```





Comparison Map
=====================================
The map below shows which side (DEM or GOP) raised more money within each individual zip code.  
The larger the circle, the greater the fundraising advantage by the winning party in that zip code.

Row {}
-----------------------------------------------------------------------

### Next trip happening

```{r}

savedmap_zipcompare

```
